Install router and use it to link to details page from main page
